// src/app/core/types/pages.ts
export enum EnumNamesPage {
  BEGIN_WORKSPACE_PAGE = "BeginWorkspacePage",
  BEGIN_PAGE = "BeginPage",
  TABS_PAGE = "TabsPage",
  WORKSPACE_PAGE = "WorkspacePage",

  PEDIDOS_PAGE = "OrdersPage",
  PEDIDO_DETALHES_PAGE = "OrderDetailsPage",
  PEDIDO_PRODUTO_DETALHES = "OrderProductDetailsPage",

  LOGIN_PAGE = "LoginPage",
  LOGIN_ALREADY_LOGGED_PAGE = "LoginAlreadyLoggedPage",

  SEARCH_PDV_PAGE = "SearchPdvPage",
  REQUEST_PAGE = "RequestsPage",
  REPORTS_PAGE = "ReportsPage",
  CONFIGS_PAGE = "ConfigsPage",
  SEARCHS_PAGE = "SearchsPage",

  LOADING_PAGE = "LoadingPage",
  SYNCHRONIZATION_PAGE = "SynchronizationPage",
  SYNCHRONIZATION_CONFIG_PAGE = "SynchronizationConfigurationsPage",
  REQUEST_TABLOIDES_PAGE = "RequestsTabloidesPage",
  REQUEST_DISTRIBUTOR_PAGE = "RequestsDistributorsPage",
  REQUEST_PRODUCTS_PAGE = "RequestsProductsPage",
  REQUEST_MIX_IDEAL_PRODUCTS_PAGE = "RequestsMixIdealProductsPage",
  REQUEST_COMBOS_PAGE = "RequestsCombosPage",
  REQUEST_FINAL_PAGE = "RequestsFinalPage",
  DETAILS_PAGE = "DetailsPage",

  PDV_DETAILS_TABS = "PdvDetailsTabsPage",
  PDV_INFORMATION_TAB = "PdvInformationTabPage",
  PDV_ADDRESS_TAB = "PdvAddressTabPage",
  PDV_CONTACT_TAB = "PdvContactTabPage",

  VISITS_AND_RESEARCHS_PAGE = "VisitsAndResearchsPage",
  VISITS_PAGE = "VisitsPage",
  VISITS_RESUME_TAB = "VisitsResumeTabPage",
  VISITS_STATUS_TAB = "VisitsStatusTabPage",
  VISITS_IMAGES_TAB = "VisitsImagesTabPage",
  VISITS_IMAGE_SUBTITLE = "ImageSubtitlePage",

  RESEARCH_WITH_STEP = "ResearchWithStepPage",
  RESEARCH_ITEM = "ResearchItemPage",
  RESEARCH_ITEM_ADD = "ResearchItemAddPage",
  RESEARCH_COMPLEMENTARY = "ResearchComplementaryPage",
  RESEARCH_TRADE = "ResearchTradePage",
  RESEARCH_TRADE_LIST_SUBMITTED = "ResearchTradeListSubmittedPage",
  RESEARCH_TRADE_SHOW_ANSWERS = "ResearchTradeShowAnswersPage",
  RESEARCH_SHARE_GONDOLA_PAGE = "ResearchShareGondolaPage",
  RESEARCH_SHARE_OF_SHELF = "ResearchShareOfShelfPage",

  RESEARCH_SHARE_GONDOLA_IMAGE = "ResearchShareGondolaImagePage",
  RESEARCH_SHARE_GONDOLA_IMAGE_MODAL = "ResearchShareGondolaImageModalPage",
  RESEARCH_SHARE_GONDOLA_IMAGE_CROP = "ResearchShareGondolaImageCropPage",

  DASHBOARD_CONFIG_PAGE = "DashboardConfigPage",

  ROUTES_PAGE = "RoutesPage",
  ROUTES_ADD_PAGE = "RoutesAddPage",

  SELECT_PDV_PAGE = "SelectPdvPage",

  ORDER_CONFIG_PAGE = "OrderConfigPage",
  APP_CONFIG_PAGE = "AppConfigPage",

  CONTENT_LISTING = "ContentListingPage",
  CONTENT_DETAILS = "ContentDetailsPage"
}